+++ CHANGELOG.md
@@ -8,7 +8,17 @@ To know more about breaking changes, see the [Migration Guide][].
 
 ## Unreleased
 
-*None.*
+### Features
+
+- Allows to get the duration of a Live Photo with `AssetEntity.durationWithOptions` on iOS and macOS.
+
+### Improvements
+
+- Improves the options when fetching fixed number of assets on iOS and macOS.
+
+### Fixes
+
+- Do not use `privateFileURL` on iOS 18+.
 
 ## 3.5.2